fix CI for now by excluding the failing assertion from running.
On the CI linux workflow, archived baseline aren't used - instead fixtures will be re-evaluated.
As of Git 2.47, its behaviour changed which makes the following assertion fail.
We decided to just ignore it until it's clear that this isn't a bug - obviously the traversal order changed.

----

This temporarily suppresses baseline comparisons for two patterns
where the baseline checks fail if the `complex_graph` fixture
script was run in Git 2.47.0.

For now, they are unconditionally suppressed, regardless of the Git
version or whether generated archives are used. This change is
specific to the `find_youngest_matching_commit::regex_matches`
test.

This works around, but does not fix, issue #1622. The affected test
is still run, but the two directly affected baseline checks are
skipped.

Co-authored-by: Eliah Kagan <degeneracypressure@gmail.com>