feat: add `tree()` and `commit()` merge support, en par with `merge-ORT` as far as tests go.
Note that this judgement of quality is based on a limited amount of partially complex
test, but it's likely that in practice there will be deviations of sorts.

Also, given the complexity of the implementation it is definitely under-tested,
but with that it's mostly en par with Git, unfortunatly.

On the bright side, some of the tests are very taxing and I'd hope this
means something for real-world quality.