Set +x in index in status_changed.sh fixture
This adds a `git update-index --chmod=+x` command in the gix-status
`status_changed.sh` fixture script before the `git commit` command,
so that the commit holds the same mode bits on Windows, where
`chmod +x` does not take effect, as on Unix-like systems where it
is generally does take effect.

This change is done for the purpose of making the fixture be as was
intended. But there is no corresponding change to be made to test
cases that use it. The tests are asserting different things on
Windows and non-Windows systems about changes in the working tree
after a commit. The fixture script deliberately does not stage such
changes, so for the `chmod +x` and `chmod -x` commands appearing
after that in the fixture script, there is no change to the index
that can correctly be made to unify the behavior across platforms.

Windows doesn't support Unix-style executable permissions in the
filesystem, so the state of the working tree still has to be
asserted differently on Windows. Although the mode is different in
the repo metadata due to this change, it still is not found to
disagree with the mode on disk, since the latter information is
absent. (Because the files are staged, is implicitly taken to agree
with whichever of +x or -x is already recorded in the repo for
them.)