Set +x in index in gix-worktree-state make_mixed* fixtures
This adds `git update-index --chmod=+x` commands to the
`make_mixed`, `make_mixed_without_submodules`, and
`make_mixed_without_submodules_and_symlinks` fixtures used in
gix-worktree-state `tests/state/checkout.rs` tests, so that files
are staged and committed with the intended modes.

The tests were actually passing before this change, but making it
verifies that the tests still work even when the repository is as
assumed. (Also, some of those tests are recently enabled or made
more expansive on Windows, due to the preceding commit that makes
them test symlinks, where applicable, on all platforms.)