Split i686 and armv7 jobs; run all tests on i686
The test failures that appeared with `i686-unknown-linux-musl` are
assertions about the size of data structures, where the actual
sizes on 32-bit targets are smaller. But these are not the only
such failing assertions in the test suite: local testing with
`i686-unknown-linux-gnu` on a 32-bit x86 Debian 12 system, and with
`i686-pc-windows-msvc` on a 32-bit x86 Windows 10 system, reveals
that there are significantly more such failures. About 20 tests
usually fail on these platforms, with most failures being of data
structure size assertions.

To catch such failures when they arise as regressions, it would be
useful to have CI coverage of most of the test suite on some 32-bit
target. Since no emulation is needed to run i686 binaries on amd64
CI runners, let try `i686-unknown-linux-musl` for this.

This doesn't include running tests with `GIX_TEST_IGNORE_ARCHIVES`,
because a test that specifically attempts to exercise fixture
scripts should run in an environment where the tools being called,
including `git`, and `bash` itself, are 32-bit builds. (That could
probably be achieved with a `container` job.)