Change i686 job to i686-unknown-linux-gnu, run in container
The previous approach did not work because musl dependencies were
not installed. (32-bit `-gnu` dependencies would also not be
installed, in that non-container environment.)

Since this is switching to a container, it may as well target the
OS of the container.

The previous reasoning for not (also?) having this 32-bit test use
`GIX_TEST_IGNORE_ARCHIVES=1` no longer applies, but that is not
done at this time.