Set nextest behavior on empty runs more explicitly
`cargo nextest` will eventually default to `--no-tests=error`, but
it doesn't yet. Currently, `cargo nextest run -p gitoxide-core`
runs no tests, as there are non in `gitoxide-core` (checked by
running `git grep '#.*\[.*test' -- gitoxide-core`).

This adds `--no-tests=warn` to that command in the `unit-tests`
recipe in `justfile`, to make the default behavior explicit.

For CI (but not otherwise), this sets the default to `fail` now by
setting the `NEXTEST_NO_TESTS` environment variable. (This is
overridden by `--no-tests=warn` so it does not fail `unit-tests`.)