Start on `announce-release` job
Some users and developers, such as downstream package maintainers,
would benefit from a way to subscribe on GitHub to notifications
for releases associated with the `gitoxide` crate, i.e. the
releases done via `release.yml`, but not of the many `gix-` crates
that they would also be notified about if they watched releases via
the GitHub watch feature.

The idea here is to create a way for people to easily be notified
by subscribing to a specific locked discussion thread that a GitHub
Actions job posts after publishing a `gitoxide` release.

I am testing this with EliahKagan#6, but that is just for testing.
If this is used, the actual announcements "discussion" thread would
be separate, and on GitoxideLabs/gitoxide (not on a fork).