Add missing executable bits on fixture scripts
Five fixture scripts were found by `etc/check-mode.sh` to begin
with `#!` but not be marked executable. This adds `+x` to them.

(I inspected them to make sure that this, and not removing the
shebangs, was the correct fix.)

This commit also removes the two temporary test files that had been
added to ensure `etc/check-mode.sh` is capable of identifying the
opposite situation. The script is now expected to pass.