Simplify the `check-mode` CI job
By running it only on Ubuntu (not macOS and Windows), and removing
parts that were only needed for running it on Windows.

Unlike `etc/copy-packetline.sh`, the `etc/check-mode.sh` script
probably does not need to be made easy to retest on other platforms
after changes, becaus it does not delete (nor otherwise modify) the
repository, and becuase it is considerably shorter, simpler, and
less reliant on subtleties of standard tools that vary across
platforms.