Fix a few ShellCheck warnings and stylistic inconsistencies
This was originally intended to be part of much broader changes, to
more scripts, along the same lines, and adding a ShellCheck CI job.
That will likely still happen, but it's not clear when it will get
done. Rather than wait, this brings in the small benefits of these
changes.