Run the `check-mode` CI job only on Ubuntu
And not on macOS on Windows.

(If it is to be kept as a separate job, then possibly caching
should be enabled for it. But I suspect it can be moved into
another existing job, most likely the `test` job, which already
runs a number of distinct checks via `just`.)