Quote `{{ it }}` in `justfile` to fix recipe in Windows
Since Windows was failing due to the shell called by `just`
treating `\` path separators as escape characters (the main effect
of which was that they were removed in quote removal).

This was not a problem for internal-tools, but it broke the `just`
recipe at least in some ways of running it on Windows, including CI
and at least one local system tested.