feat!: add support for not tracking single empty files.
Empty files are equivalent to having no content, which also means
such files have no identity to speak off. This definitely helps
with false positives of `.gitignore` for instance, which can be empty
to tell Git to track a directory.

On top of that, Git has a heuristic to do rename tracking of small files
by similarity as the similarity may be off of files just have a couple of
lines to speak about.

Note that empty files that are renamed as part of a whole directory will still
be tracked as renames.