feat: add `topo::Builder::with_ends` for adding ends after from_iters
Currently, `Builder::from_iters` takes as arguments both the tips and
ends, and it's the only possibility for specifying either of them during
the building process. Considering that `ends` in `Builder::from_iters`
is an `Option`, this is not very builder-like and obstructs some
use-cases.

This change introduces a fn which allows adding additional ends after
initial construction of a "fresh" `Builder`.