fix: Don't attempt +x if it was replaced by a non-regular file
This does not make a difference in typical cases, and anytime it
matters something has probably gone unexpectedly, but this narrows
the window of time during which a race condition could occur where
a regular file has been replaced with something else at the same
path (e.g. a directory) by some other process.

An example of why it's valuable to avoid this is that if the entry
is a directory then executable permissions have a different meaning
and adding them could increase access unintentionally. Likewise,
when we set executable permissions we remove setuid, setgid, and
sticky bits, which also have different meanings for directories; in
particular, on a directory the sticky bit restricts deletion.

(This also automatically avoids some problems in the case of
`finalize_entry` being called with a path to set executable that
was never a regular file in the first place. That should not
happen, though, and allowing that is not a goal of this change.)