Test for excessive write permissions on nonexclusive checkout
This expands the `overwriting_files_and_lone_directories_works`
test function to check not only that executable permissions have
been added at least for the owner, but that write permissions are
not added anywhere they would be excessive.

The write permissions bits of the current test runner process umask
is used to gauge this. This means we are expecting different
permissions in different environments.

`(umask ...; cargo nextest run ...)` can be used to try it with
different umasks if desired, but this should not typically be
necessary.

Possible under-restrictive umasks could make the test pass even if
the underlying bug is not fixed; this is avoided by also testing
that the umask is sufficient to facilitate the test.

(This is really why the test accesses the umask in the first place:
in environments where files would automatically be created with
completely unrestricted permissions, the expected behavior of the
code under test may be to do that, but running the tests in such an
environment is insufficient to check if the bug is fixed.)