Fix an MSRV incompatibility
Trimming Unicode whitespace is not equivalent to trimming ASCII
whitespace, but in this case I think they are equally okay to do.

If we get Unicode whitespace that is not ASCII whitespace from
running `umask`, in principle it's possible this is due to it being
a misinterpretation of something not Unicode. But if `umask` gives
us anything besides whitespace and a single nonempty sequence of
octal digits, then (a) that's very strange, and (b) it should fail
and indicate its failure with a nonzero exit status, which we
check.