Use the ARM runner for some 32-bit tests
This turns the `test-32bit` job into a matrix job definition with
two jobs: the i386 job, which already existed, and a new ARMv7 job,
which is able to run (still without emulation) in a container on
the new ARM64 runner.

As written, this is not expected to work, and is even expected to
break the old job, because I've commented out the code that
installs libraries for the 64-bit host architecture inside the
container to allow the 64-bit Node.js on the host to run in the
container for script steps. This is just to make sure that is still
needed, before attempting to generalize it to cover both platforms.