Convert the mode more portably
The previous conversion worked on Linux but failed on macOS where
modes, as represented in system data structures, are 16-bit rather
than 32-bit.

The previous commit broke builds on macOS (and possibly other
OSes not currently tested on CI), which this should fix.

This commit also does some other refactorings:

- Simplify the way conversions are represented.

- Express with `expect` that, by this point, there are no unknown
  bits, rather than doing something that would silently preserve or
  silently discard them. There would have to be a bug (and probably
  in `gix-worktree-state` itself) for that not to be the case.

- Clarify the TODO comment, and also weaken it since there might be
  some reason for `set_mode_executable` to keep using `Permissions`
  in some way.