Draft experiment to investigate `rustc` SIGSEGV
This is to investigate the problem on the `test-fast` job with the
new ARM64 runner described in #1790.

This experiment does not produce useful results yet, because it has
no way to distinguish happenstance from correlation. To do that, I
need either to rerun each job repeatedly, or further parameterize
the matrix to do that. I'll be doing the latter, but right now this
dimension has size 1 (i.e., the only value of `number` is `0`)
so I don't start a large number of jobs when something is broken
due to a mistake in the workflows.