Make `test-fixtures-windows` required for PR auto-merge
Since #1663, the `test-fixtures-windows` CI job checks actual
failures against a list of specific tests that are known to fail on
Windows when `GIX_TEST_IGNORE_ARCHIVES=1`. It is therefore capable
of providing useful information about new failures, or newly
passing tests that should be removed from the list, if the job ever
does fail.

The job also seems not to fail. This is to say that while #1358 is
not fixed, the `test-fixtures-windows` job has a very low rate of
failure and, if it does fail, something new and interesting would
be happening such that we would want to know about it and probably
not immediately merge a PR that caused it without checking how and
why that happened.

This adds `test-fixtures-windows` to the list of jobs that are
dependencies of a required check for branch protection based PR
auto-merge.