Draft experiment 3 to try more ways to install
In case the installation method makes a difference.

Also, this brings back testing of the unstable toolchain.

This has just one job for each meaningful combination, so mistakes
in the experiment workflow can be found before doing nine times
as much work. The experiment this prepares should hopefully shed
more light on #1790 (or increase confidence in the observations so
far), but this is just preparation: variation across runs will
likely be due to the bug being nondeterministic.