Test `let_readers_execute` on all targets
The `let_readers_execute` function computes a new mode with `+x`
set reasonably based on an old mode. It is only useful on Unix-like
systems. But there may be a benefit to being able to detect more
bugs affecting one platform even on another; this can help catch
bugs slightly earlier in the development process.

Now that `let_readers_execute` only longer depends on a type that
varies across Unix-like systems and is only defined on such systems
(instead representing modes as `u32` on all of them), it no longer
contains anything that can't build or be tested everywhere. So this
lets its tests run on all systems, even non-Unix-like ones.