Use `br` literals to express `\` paths more clearly in tests
In `gix-validate` tests that Windows-style paths are validated
correctly, this changes byte strings that contain `\` to be raw
byte string literals (`rb`) instead of non-raw byte string literals
(`b`).

This is to improve readability given that some literals contain
consecutive backslashes: this makes it easier to tell if, and how
many, consecutive backslashes there are.